requirements.txt
langchain

langchain-aws

langgraph

boto3

python-dotenv

streamlit

pydantic